Move generic encryption methods to encryptor_service.py
Some code shared by gce and aws (and likely other csps)
was in encrypt_ami.py. It makes more sense for this code
to be in the generic encryption service file.

Testing: encrypted and updated gce images
